* Added BaseStationTime_t to MsgPacket
* Changed return type of GetNumPendingMsgPackets

* Fixed namespace scope of IComms in BasestationMain

* Added json config as param to BasestationMain. (Not used yet)

+++ basestation/src/uiTcpComms.cpp
@@ -194,9 +194,9 @@ namespace Cozmo {
   }
   
   
-  size_t UiTCPComms::GetNumPendingMsgPackets()
+  u32 UiTCPComms::GetNumPendingMsgPackets()
   {
-    return recvdMsgPackets_.size();
+    return (u32)recvdMsgPackets_.size();
   };
   
   void UiTCPComms::ClearMsgPackets()
